Added more methods.., classes...
Continuation of the backend development. Added collision detection for Ball to Ball and Ball to Table. Filled other classes and implemented more functions.

To test your methods use app.post() function. This is done to simulate what it would be like when you would launch the game.. as well as allowing you to test the HTML pages without imidietly launching the whole game.